requirements.txt
build>=0.6.0
twine>=3.4.2
spade==3.2.3
python-telegram-bot==20.1